More modular files added. Lots of code cleaning done to main index page.
There was some strange code at the end of all of the files. I moved that
into a seperate file named end-stuff for now. I will clean that more
later if i get the chance.

Code cleaning on the index page mainly consisted of clearing out a lot
of unsed attributes for images. I think this was originaly designed in
wordpress. Added some basic comments.