Updated most of the links to relative. Minor changes to publications page
All the links that make sense to be relative are now relative. This will
help with website redeployment.

Two changes to the publications page. First is move the capstone section
above the invited talk section. Next remove Kelvins name from the
invited talks and replace it with the location of the talk.